+++ CHANGES.md
@@ -6,6 +6,7 @@
   now relative to ocaml parent directory so all `git am` based patches are of form
   `flexdll/flexdll.c`, etc. to make robust to the lack of `flexdll/.git`.
    Vendor supplied patches are used by DkSDK CMake's `110-ocaml-lang`.
+* Supply `DKML_HOST_ABI` to post-transform configure scripts.
 
 ## 2.1.3